[FIX] Lib my_free_arrays moved to the array part
