+++ lib/my/include/my.h
@@ -20,4 +20,3 @@ static const char * const HEXA_UPPER_BASE = "0123456789ABCDEF";
 
 void my_free(int size, ...);
 void *my_free_null(int size, ...);
-void my_free_arrays(int size, ...);
