+++ lib/my/include/my_arrays.h
@@ -9,3 +9,4 @@
 
 char **my_arraydup(char **array);
 int my_arraylen(char * const *array);
+void my_free_arrays(int size, ...);
