+++ src/command/output.c
@@ -10,7 +10,7 @@
 #include <stdio.h>
 #include "messages.h"
 #include "my_string.h"
-#include "my.h"
+#include "my_arrays.h"
 #include "redirections.h"
 
 static bool check_null_ambiguous(command_t *command, char **array, int index)
